Upgrade rubocop to the latest version
Since we have been using reevoocop the the underlying version of rubocop has become outdated.

The advantages of moving to a newer rubocop are:
* More cops to choose from
* Speed - (on mark the speedup is 1.95s vs 22s to lint the full codebase)
* We can loose the monkeypatch on the Rake task

Clearly the rules that ship with rubocop have changed somewhat so I have
used the mark repo as a test suite to see what is involved, this
is because with the existing version of reevoocop mark passes on all
files.

There are 3 categories of stuff to consider:

* New cops and other changes that seem sensible and will autocorrect with `reevoocop -a` see https://github.com/reevoo/mark/commit/2f86a8361673dd7c530353c6b56d423dbdd43446 for an example
* New cops and changes that cannot be autocorrected ... see the changes to lib/reevoocop.yml for these
* Lint errors that indicate incorrect or ambiguous ruby code, these need to be fixed manualy, but we should do so https://github.com/reevoo/mark/commit/36e9fad83aa8c30991d82289ff0d0a937745e5c2